+++ src/models/search.rs
@@ -67,8 +67,8 @@ pub struct SearchResult {
 #[serde(rename_all = "PascalCase")]
 pub struct Pagination {
   pub page: u64,
-  pub page_next: Option<usize>,
-  pub page_prev: Option<usize>,
+  pub page_next: Option<u64>,
+  pub page_prev: Option<u64>,
   pub page_total: u64,
   pub results: u64,
   pub results_per_page: u64,
